Module Picasso.Rendering

Module for 2d drawing of abstract elements, handles the 'camera' settings, 2D projection, and some graphical options

type t

Type of 2D scenes

val create : ?title:string -> ?padding:float -> ?grid:bool -> ?axis:bool -> abciss:string -> ordinate:string -> float -> float -> t

Initalizes an empty 2d scene.

val add : ?autofit:bool -> t -> (Colors.t * Drawable.t) -> t

Registers an abstract element, associated to a color, into a scene. Automatically changes the camera settings to encompass the newly added abstract element. You can cancel this behaviour by settings the optional argument autofit to false

val add_l : ?autofit:bool -> t -> (Colors.t * Drawable.t) list -> t

Registers a list of element into a scene. add_l r l is equivalent to calling add successively on the elements of l

Camera settings

val translate : (float * float) -> t -> t

translation of the whole scene

val scale : t -> float -> t

zoom / unzoom